Move splitter to module_utils so modules can use it and fix command arg splitting
Fixes #8338
